+++ projects/komponentkartan/src/lib/controls/icon/icon.module.ts
@@ -1,5 +1,4 @@
-import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { library } from '@fortawesome/fontawesome-svg-core';
+import { FontAwesomeModule, FaIconLibrary } from '@fortawesome/angular-fontawesome';
 import { fas } from '@fortawesome/free-solid-svg-icons';
 import { far } from '@fortawesome/free-regular-svg-icons';
 import { NgModule } from '@angular/core';
@@ -11,8 +10,8 @@ import { NgModule } from '@angular/core';
 })
 
 export class IconModule {
-    constructor() {
+    constructor(library: FaIconLibrary) {
         // Add an icon to the library for convenient access in other components
-        library.add(fas, far);
+        library.addIconPacks(fas, far);
     }
 }