+++ projects/komponentkartan/src/lib/controls/list/listComponent.angular.spec.ts
@@ -1,6 +1,6 @@
 import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DebugElement, Renderer, ElementRef, Input, Component } from '@angular/core';
+import { DebugElement, ElementRef, Input, Component, Renderer2 } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 
@@ -59,7 +59,7 @@ describe('[ListComponent - Angular]', () => {
       imports: [CommonModule, BrowserAnimationsModule, FontAwesomeModule, IconModule],
       providers: [
         { provide: ElementRef },
-        { provide: Renderer },
+        { provide: Renderer2 },
         ListService
       ]
     });